==> 1_Data_Preprocessing.Rmd <==
---
title: "Data Preprocessing"
author: "Joy Fu"
date: "2025-05-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
rm(list = ls())
pacman::p_load(tidyverse, PheWAS, gtsummary)
raw_data_path = "/Users/joyfu/Desktop/Projects/AD-LOE-genetics/data/"
output_path = "/Users/joyfu/Desktop/Projects/AD-LOE-genetics/output/"
'%!in%' = function(x,y)!('%in%'(x,y))
```

# Part 0. Prepare for SQL extraction
```{r}
epi_phecode = c("345", "345.1", "345.11", "345.12", "345.3")
ad_phecode = "290.11"
covariate_phecode = c("401.1", "250.2", "433", "272.1")

icd_of_interest = PheWAS::phecode_map %>% filter(vocabulary_id == "ICD10CM") %>% 
  filter(phecode %in% c(epi_phecode, ad_phecode, covariate_phecode)) %>%
  pull(code) %>% unique() %>% sort()
length(icd_of_interest) # 118
# paste the above icd codes into a string with single quotes
icd_of_interest_str = paste0("'", icd_of_interest, "'", collapse = ", ")
# icd_of_interest_str
```

```{r}
# Epilepsy phecodes only
icd_of_interest = PheWAS::phecode_map %>% filter(vocabulary_id == "ICD10CM") %>% 
  filter(phecode %in% c(epi_phecode)) %>%
  pull(code) %>% unique() %>% sort()
length(icd_of_interest) # 118
# paste the above icd codes into a string with single quotes
icd_of_interest_str = paste0("'", icd_of_interest, "'", collapse = ", ")
icd_of_interest_str
```


# Part 1. DDR Data Cleaning
## 1. Clean encounters info (with covariates)
```{r, eval=FALSE}
phe_table_extract_date = "2025-05-21"
# Read in DDR patient data
ddr_case_raw = read_csv(file = paste0(raw_data_path, 
                                      "ddr/raw/ddr_over60_case_20250521.csv"),
                        col_names = FALSE, na = c("NULL"))
names(ddr_case_raw) = c("PatientID", "Sex", "Ethnicity", "FirstRace", 
                        "BirthDate", "DeathDate", "PatientLivingStatus", 
                        "DateFirst", "DateLast", "DiagnosisDate", "DiagnosisCode")
dim(ddr_case_raw) # dim = (818116,11)
# Clean encounter info
ddr_case_clean = ddr_case_raw %>% 
  mutate(BirthDate = as.Date(BirthDate), DeathDate = as.Date(DeathDate),
         DateFirst = as.Date(DateFirst), DateLast = as.Date(DateLast),
         DiagnosisDate = as.Date(DiagnosisDate)) %>% 
  mutate(valid_record = case_when(
    is.na(DeathDate) & DiagnosisDate >= BirthDate ~ 1,
    !is.na(DeathDate) & DiagnosisDate >= BirthDate & DiagnosisDate <= DeathDate ~ 1
  )) %>% filter(valid_record == 1) %>% select(-valid_record) %>% 
  filter(Sex != "X" & Sex != "Unknown")
dim(ddr_case_clean) # dim = (817877,11)
ddr_case_demo = ddr_case_clean %>% 
  select(PatientID, Sex, Ethnicity, FirstRace, BirthDate, DeathDate, DateFirst, 
         DateLast) %>% unique() %>% 
  mutate(age_first_visit = as.numeric(DateFirst - BirthDate)/365.25) %>%
  mutate(age_last_visit = as.numeric(DateLast - BirthDate)/365.25) %>% 
  mutate(record_length = as.numeric(DateLast - DateFirst)/365.25)
dim(ddr_case_demo) # dim = (395984,11)
```

```{r}
# Merge with phecode
ddr_case_enc_phecode = ddr_case_clean %>% mutate(vocabulary_id = "ICD10CM") %>% 
  select(PatientID, DiagnosisDate, DiagnosisCode, vocabulary_id) %>% 
  unique() %>% arrange(PatientID, DiagnosisDate) %>%
  left_join(PheWAS::phecode_map, by = c("vocabulary_id" = "vocabulary_id",
                                        "DiagnosisCode" = "code")) %>% 
  unique() %>% select(-DiagnosisCode) %>% unique() %>% filter(!is.na(phecode)) 
dim(ddr_case_enc_phecode) # dim = (798824,4)
save(ddr_case_enc_phecode, file = paste0(raw_data_path, "ddr/mod/ddr_case_enc_phecode.RData"))
```

## 2. Define cases
```{r}
# Select epilepsy case and controls
epi_phecode = c("345", "345.1", "345.11", "345.12", "345.3")
all_epi_case = ddr_case_enc_phecode %>% 
  filter(phecode %in% epi_phecode) %>% pull(PatientID) %>% unique()
length(all_epi_case) # N = 16273
LOE_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% epi_phecode) %>% left_join(ddr_case_demo) %>%
  mutate(age_diagnosis = as.numeric(DiagnosisDate - BirthDate)/365.25) %>%
  select(PatientID, DiagnosisDate, phecode, age_diagnosis, age_first_visit) %>% 
  group_by(PatientID) %>% 
  mutate(first_epi_age = min(age_diagnosis), 
         first_EpilepsyDate = min(DiagnosisDate)) %>% 
  select(PatientID, first_EpilepsyDate, first_epi_age, age_first_visit) %>% 
  filter(first_epi_age >= 60 & age_first_visit < 60) %>% 
  unique() %>% left_join(ddr_case_demo) %>% mutate(epilepsy = 1)
dim(LOE_case_info) # dim = (4973,14)
rm_LOE_id = setdiff(all_epi_case, LOE_case_info$PatientID)
length(rm_LOE_id) # 11300
```

```{r, eval=FALSE}
# Select AD case and controls
ad_phecode = "290.11"
all_AD_case = ddr_case_enc_phecode %>% 
  filter(phecode %in% ad_phecode) %>% pull(PatientID) %>% unique()
length(all_AD_case) # N = 11700
AD_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% ad_phecode) %>% left_join(ddr_case_demo) %>%
  mutate(age_diagnosis = as.numeric(DiagnosisDate - BirthDate)/365.25) %>%
  select(PatientID, DiagnosisDate, phecode, age_diagnosis) %>% group_by(PatientID) %>% 
  mutate(first_AD_age = min(age_diagnosis), first_ADDate = min(DiagnosisDate)) %>% 
  select(PatientID, first_ADDate, first_AD_age) %>% 
  unique() %>% left_join(ddr_case_demo) %>% mutate(AD = 1)
dim(AD_case_info) # dim = (11700,14)
```

```{r}
# Other covariates
# HTN
htn_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% c("401.1")) %>% group_by(PatientID) %>% 
  mutate(HTNDate = min(DiagnosisDate), HTN = 1) %>% ungroup() %>%
  select(PatientID, HTN, HTNDate) %>% unique()
dim(htn_case_info) # dim = (305256,3)
# Diabetes
diab_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% c("250.2")) %>% group_by(PatientID) %>% 
  mutate(DiabetesDate = min(DiagnosisDate), diabetes = 1) %>% ungroup() %>%
  select(PatientID, diabetes, DiabetesDate) %>% unique()
dim(diab_case_info) # dim = (117539,3)
# Stroke
stroke_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% c("433")) %>% group_by(PatientID) %>% 
  mutate(StrokeDate = min(DiagnosisDate), stroke = 1) %>% ungroup() %>%
  select(PatientID, stroke, StrokeDate) %>% unique()
dim(stroke_case_info) # dim = (5039,3)
# Hyperlipid
hyperlipid_case_info = ddr_case_enc_phecode %>% 
  filter(phecode %in% c("272.1")) %>% group_by(PatientID) %>% 
  mutate(HyperlipidDate = min(DiagnosisDate), hyperlipid = 1) %>% ungroup() %>%
  select(PatientID, hyperlipid, HyperlipidDate) %>% unique()
dim(hyperlipid_case_info) # dim = (248025,3)
```

```{r}
case_final = ddr_case_demo %>% 
  left_join(LOE_case_info %>% select(PatientID, first_EpilepsyDate, first_epi_age)) %>%
  left_join(AD_case_info %>% select(PatientID, first_ADDate, first_AD_age)) %>% 
  mutate(AD = if_else(PatientID %in% all_AD_case, 1, 0),
         LOE = if_else(PatientID %in% LOE_case_info$PatientID, 1, 0)) %>% 
  left_join(htn_case_info) %>% left_join(diab_case_info) %>% 
  left_join(stroke_case_info) %>% left_join(hyperlipid_case_info) %>% 
  mutate(HTN = if_else(!is.na(HTN), 1, 0),
         diabetes = if_else(!is.na(diabetes), 1, 0),
         stroke = if_else(!is.na(stroke), 1, 0),
         hyperlipid = if_else(!is.na(hyperlipid), 1, 0)) %>% unique() %>% 
  mutate(female = if_else(Sex == "Female", 1, 0)) %>% select(-Sex) %>% 
  mutate(deceased = if_else(!is.na(DeathDate), 1, 0)) %>% 
  filter(PatientID %!in% rm_LOE_id)
dim(case_final) # dim = (384684, 26)
# Select only the patients in NH-White
white_group = c("White", "White or Caucasian", "European", "White: Not Listed")
ddr_case_NHWhite = case_final %>% 
  filter(Ethnicity == "Not Hispanic or Latino" & FirstRace %in% white_group) %>% 
  select(PatientID, female, deceased, BirthDate, DeathDate, age_first_visit, 
         age_last_visit, record_length, first_EpilepsyDate, first_epi_age, 
         first_ADDate, first_AD_age, AD, LOE, HTN, diabetes, stroke, hyperlipid) %>% 
  unique()
dim(ddr_case_NHWhite) # dim = (192614, 18)
```

## 3. Define controls
```{r}
# Read in DDR patient data
ddr_control_raw = read_csv(file = paste0(raw_data_path, 
                                      "ddr/raw/ddr_over60_control_20250521.csv"),
                        col_names = FALSE, na = c("NULL"))
names(ddr_control_raw) = c("PatientID", "Sex", "Ethnicity", "FirstRace", 
                        "BirthDate", "DeathDate", "PatientLivingStatus", 
                        "DateFirst", "DateLast")
dim(ddr_control_raw) # dim = (886320,9)
# Clean demographic info
ddr_control_demo = ddr_control_raw %>% 
  filter(PatientID %!in% ddr_case_raw$PatientID) %>%
  mutate(BirthDate = as.Date(BirthDate), DeathDate = as.Date(DeathDate),
         DateFirst = as.Date(DateFirst), DateLast = as.Date(DateLast)) %>% 
  filter(Sex != "X" & Sex != "Unknown") %>% 
  select(PatientID, Sex, Ethnicity, FirstRace, BirthDate, DeathDate, DateFirst, 
         DateLast) %>% unique() %>% 
  mutate(age_first_visit = as.numeric(DateFirst - BirthDate)/365.25) %>%
  mutate(age_last_visit = as.numeric(DateLast - BirthDate)/365.25) %>% 
  mutate(record_length = as.numeric(DateLast - DateFirst)/365.25) 
dim(ddr_control_demo) # dim = (488374,11)

ddr_control_NHWhite = ddr_control_demo %>% 
  mutate(female = if_else(Sex == "Female", 1, 0)) %>% 
  mutate(deceased = if_else(!is.na(DeathDate), 1, 0)) %>% 
  mutate(AD = 0, LOE = 0, HTN = 0, diabetes = 0, stroke = 0, hyperlipid = 0) %>%
  mutate(first_EpilepsyDate = NA, first_epi_age = NA, 
         first_ADDate = NA, first_AD_age = NA) %>%
  filter(Ethnicity == "Not Hispanic or Latino" & FirstRace %in% white_group) %>% 
  select(PatientID, female, deceased, BirthDate, DeathDate, age_first_visit, 
         age_last_visit, record_length, first_EpilepsyDate, first_epi_age, 
         first_ADDate, first_AD_age, AD, LOE, HTN, diabetes, stroke, hyperlipid) %>% 
  unique() %>% filter(record_length > 0 & age_last_visit >= 60 & age_first_visit < 90 & age_first_visit < 60)
dim(ddr_control_NHWhite) # dim = (103295, 18)
```

## 4. Combine cases and controls
```{r}
length(intersect(ddr_case_NHWhite$PatientID, ddr_control_NHWhite$PatientID)) # 0
# Combine cases and controls
ddr_full = rbind(ddr_case_NHWhite, ddr_control_NHWhite) %>% as.data.frame() %>%
  arrange(PatientID) %>% unique() %>% 
  filter(record_length > 0 & age_last_visit >= 60 & age_first_visit < 90)
dim(ddr_full) # dim = (293472, 18)
length(unique(ddr_full$PatientID)) # 293472
save(ddr_full, file = paste0(raw_data_path, "ddr/mod/ddr_full.RData"))
```

# Part 2. Subset to ATLAS
```{r}
rm(list = ls())
pacman::p_load(tidyverse, PheWAS)
raw_data_path = "/Users/joyfu/Desktop/Projects/AD-LOE-genetics/data/"
output_path = "/Users/joyfu/Desktop/Projects/AD-LOE-genetics/output/"
'%!in%' = function(x,y)!('%in%'(x,y))
phe_table_extract_date = "2025-05-20"
load(file = paste0(raw_data_path, "ddr/mod/ddr_full.RData"))
```

## 1. Clean encounters info
```{r, eval=FALSE}
# Read in ATLAS patient data
atlas_full_raw = read_csv(file = paste0(raw_data_path, 
                                        "atlas/pheno/raw/biobank_all_20250520.csv"),
                          col_names = FALSE, na = c("NULL"))
names(atlas_full_raw) = c("PatientID", "UniqueSampleId", "IsExcluded", "Sex", 
                          "Ethnicity", "FirstRace", "BirthDate", "DeathDate", 
                          "PatientLivingStatus", "DateFirst", "DateLast", 
                          "DiagnosisDate", "DiagnosisCode")
dim(atlas_full_raw) # dim = (22575586,13)
# Clean encounter info
atlas_full = atlas_full_raw %>% 
  mutate(BirthDate = as.Date(BirthDate), DeathDate = as.Date(DeathDate),
         DateFirst = as.Date(DateFirst), DateLast = as.Date(DateLast),
         DiagnosisDate = as.Date(DiagnosisDate)) %>% 
  mutate(valid_record = case_when(
    is.na(DeathDate) & DiagnosisDate >= BirthDate ~ 1,
    !is.na(DeathDate) & DiagnosisDate >= BirthDate & DiagnosisDate <= DeathDate ~ 1
  )) %>% filter(valid_record == 1) %>% select(-valid_record) %>% 
  filter(Sex != "X" & Sex != "Unknown")
dim(atlas_full) # dim = (22573686,13)
rm(atlas_full_raw)
```

## 2. Clean patients info
```{r}
atlas_data_path = "/Users/joyfu/Desktop/Projects/Dementia-prediction/data/"
load(file = paste0(atlas_data_path, "atlas/geno/gen_ancestry.rda"))
# Merge patient data with unique sample ID
enc_demo = atlas_full %>% filter(!is.na(UniqueSampleId)) %>% 
  filter(is.na(IsExcluded)) %>% select(-IsExcluded) %>% 
  select(PatientID, UniqueSampleId, Sex, BirthDate, DeathDate, DateFirst, 
         DateLast) %>% unique() 
dim(enc_demo) # dim = (57648,7)

# Clean patients info
enc_demo_clean = enc_demo %>% 
  filter(UniqueSampleId %in% gen_ancestry$UniqueSampleId) 
enc_demo_clean %>% group_by(PatientID) %>%
  summarize(n = n()) %>% filter(n > 1) %>% nrow() # 0
enc_demo_clean = enc_demo_clean %>%
  mutate(age_first_visit = as.numeric(DateFirst - BirthDate)/365.25) %>%
  mutate(age_last_visit = as.numeric(DateLast - BirthDate)/365.25) %>% 
  mutate(record_length = as.numeric(DateLast - DateFirst)/365.25) %>% 
  inner_join(gen_ancestry) %>% filter(!is.na(gen_ancestry))
dim(enc_demo_clean) # dim = (27253,11)
save(enc_demo_clean, file = paste0(raw_data_path, "atlas/pheno/mod/enc_demo_clean.rda"))
```

## 3. Finalize patients records and demo info
```{r, eval=FALSE}
# Add PCs and APOE
geno_path = "/Users/joyfu/Desktop/Projects/Dementia-prediction/data/atlas/geno/"
# Add genetic info
load(paste0(geno_path, "apoe.rda"))
load(paste0(geno_path, "pcs/anc_pcs.rda"))
load(file = paste0(raw_data_path, "prs/mod/AD_PRS_final.rda"))
apoe$UniqueSampleId = as.numeric(apoe$UniqueSampleId)
anc_pcs$UniqueSampleId = as.numeric(anc_pcs$UniqueSampleId)

atlas_genetic_short = enc_demo_clean %>% 
  select(PatientID, UniqueSampleId, gen_ancestry) %>% unique()
atlas_subset = ddr_full %>% inner_join(atlas_genetic_short) %>% 
  left_join(apoe) %>% left_join(anc_pcs) %>% left_join(AD_PRS_final) %>% 
  filter(!is.na(AD_PRS_rmapoe_norm))
# check concordance of self-reported & gen_ancestry
table(atlas_subset$gen_ancestry) # all European
dim(atlas_subset) # dim = (16500, 45)
save(atlas_subset, file = paste0(raw_data_path, "atlas/pheno/final/atlas_subset.rda"))
```


# Part 3. Descriptive statistics
## 1. DDR (by LOE/AD)
```{r, eval=FALSE}
# Desc by epilepsy status
ddr_full %>% mutate(LOE = as.factor(LOE)) %>% 
  select(LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid) %>% 
  tbl_summary(by = LOE, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```

```{r}
# Desc by epilepsy status
ddr_full %>% mutate(LOE = as.factor(LOE)) %>% 
  select(LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid) %>% 
  tbl_summary(by = LOE, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```



```{r, eval=FALSE}
# Desc by AD status
ddr_full %>% mutate(AD = as.factor(AD)) %>% 
  select(AD, LOE, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid) %>% 
  tbl_summary(by = AD, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```

## 2. ATLAS vs. non-ATLAS
```{r}
ddr_non_atlas = ddr_full %>% 
  filter(PatientID %!in% atlas_subset$PatientID) %>%
  mutate(group = "non-ATLAS") %>% 
  select(PatientID, LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, group) %>% unique()
dim(ddr_non_atlas) # dim = (399712, 13)
ddr_atlas = atlas_subset %>% mutate(group = "ATLAS") %>%
  select(all_of(names(ddr_non_atlas)))  %>% unique()
dim(ddr_atlas) # dim = (16500, 13)
atlas_compare = rbind(ddr_non_atlas, ddr_atlas) %>% as.data.frame()
dim(atlas_compare) # dim = (416212, 13)
```

```{r}
# Desc by group status
atlas_compare %>% 
  select(LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, group) %>% 
  tbl_summary(by = group, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p() %>% add_overall() 
```

## 3. ATLAS (by LOE/AD)
```{r, eval=FALSE}
# Desc by epilepsy status
atlas_subset %>% mutate(LOE = as.factor(LOE)) %>% 
  select(LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, e4count, AD_PRS_rmapoe_norm) %>% 
  tbl_summary(by = LOE, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```

```{r}
# Desc by epilepsy status
atlas_subset %>% mutate(LOE = as.factor(LOE)) %>% 
  select(LOE, AD, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, e4count, AD_PRS_rmapoe_norm) %>% 
  tbl_summary(by = LOE, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```


```{r, eval=FALSE}
# Desc by AD status
atlas_subset %>% mutate(AD = as.factor(AD)) %>% 
  select(AD, LOE, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, e4count, AD_PRS_rmapoe_norm) %>% 
  tbl_summary(by = AD, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```


```{r}
# Desc by AD status
atlas_subset %>% mutate(AD = as.factor(AD)) %>% 
  select(AD, LOE, age_first_visit, age_last_visit, female, record_length, 
         deceased, HTN, diabetes, stroke, hyperlipid, e4count, AD_PRS_rmapoe_norm) %>% 
  tbl_summary(by = AD, 
              statistic = list(all_continuous() ~ "{mean} ({sd})"),
              digits = all_continuous() ~ 2) %>% add_p()
```
